refactor: rename package from apiServece to apiService for consistency
